move kotlin into its own module
