+++ src/gen_kotlin/templates/Enum.kt
