+++ src/gen_kotlin/templates/Record.kt
