+++ src/gen_kotlin/templates/macros.kt
